+++ Debug/Core/Src/main.d
@@ -35,11 +35,12 @@ Core/Src/main.o: ../Core/Src/main.c ../Core/Inc/main.h \
  ../Drivers/STM32L4xx_HAL_Driver/Inc/stm32l4xx_hal_spi_ex.h \
  ../Drivers/STM32L4xx_HAL_Driver/Inc/stm32l4xx_hal_uart.h \
  ../Drivers/STM32L4xx_HAL_Driver/Inc/stm32l4xx_hal_uart_ex.h \
- ../Core/Inc/i2c.h ../Core/Inc/main.h ../Core/Inc/iwdg.h \
+ ../Core/Inc/iwdg.h ../Core/Inc/main.h \
+ ../Core/Inc/stm32l476g_discovery_glass_lcd.h \
+ ../Core/Inc/stm32l476g_discovery.h ../Core/Inc/i2c.h ../Core/Inc/iwdg.h \
  ../Core/Inc/lcd.h ../Core/Inc/quadspi.h ../Core/Inc/rtc.h \
  ../Core/Inc/sai.h ../Core/Inc/spi.h ../Core/Inc/usart.h \
- ../Core/Inc/gpio.h ../Core/Inc/stm32l476g_discovery_glass_lcd.h \
- ../Core/Inc/stm32l476g_discovery.h
+ ../Core/Inc/gpio.h
 ../Core/Inc/main.h:
 ../Drivers/STM32L4xx_HAL_Driver/Inc/stm32l4xx_hal.h:
 ../Core/Inc/stm32l4xx_hal_conf.h:
@@ -77,8 +78,11 @@ Core/Src/main.o: ../Core/Src/main.c ../Core/Inc/main.h \
 ../Drivers/STM32L4xx_HAL_Driver/Inc/stm32l4xx_hal_spi_ex.h:
 ../Drivers/STM32L4xx_HAL_Driver/Inc/stm32l4xx_hal_uart.h:
 ../Drivers/STM32L4xx_HAL_Driver/Inc/stm32l4xx_hal_uart_ex.h:
-../Core/Inc/i2c.h:
+../Core/Inc/iwdg.h:
 ../Core/Inc/main.h:
+../Core/Inc/stm32l476g_discovery_glass_lcd.h:
+../Core/Inc/stm32l476g_discovery.h:
+../Core/Inc/i2c.h:
 ../Core/Inc/iwdg.h:
 ../Core/Inc/lcd.h:
 ../Core/Inc/quadspi.h:
@@ -87,5 +91,3 @@ Core/Src/main.o: ../Core/Src/main.c ../Core/Inc/main.h \
 ../Core/Inc/spi.h:
 ../Core/Inc/usart.h:
 ../Core/Inc/gpio.h:
-../Core/Inc/stm32l476g_discovery_glass_lcd.h:
-../Core/Inc/stm32l476g_discovery.h:
