+++ Debug/Core/Src/main.su
@@ -1,7 +1,4 @@
-../Core/Src/main.c:74:5:main	8	static
-../Core/Src/main.c:134:6:SystemClock_Config	96	static
-../Core/Src/main.c:191:6:PeriphCommonClock_Config	144	static
-../Core/Src/main.c:213:6:LCDBarDemo	24	static
-../Core/Src/main.c:238:6:LCDCharDemo	24	static
-../Core/Src/main.c:266:6:LCDStringDemo	80	static
-../Core/Src/main.c:305:6:Error_Handler	4	static,ignoring_inline_asm
+../Core/Src/main.c:70:5:main	8	static
+../Core/Src/main.c:130:6:SystemClock_Config	96	static
+../Core/Src/main.c:187:6:PeriphCommonClock_Config	144	static
+../Core/Src/main.c:216:6:Error_Handler	4	static,ignoring_inline_asm
