+++ Debug/Core/Src/sai.d
@@ -34,7 +34,9 @@ Core/Src/sai.o: ../Core/Src/sai.c ../Core/Inc/sai.h ../Core/Inc/main.h \
  ../Drivers/STM32L4xx_HAL_Driver/Inc/stm32l4xx_hal_spi.h \
  ../Drivers/STM32L4xx_HAL_Driver/Inc/stm32l4xx_hal_spi_ex.h \
  ../Drivers/STM32L4xx_HAL_Driver/Inc/stm32l4xx_hal_uart.h \
- ../Drivers/STM32L4xx_HAL_Driver/Inc/stm32l4xx_hal_uart_ex.h
+ ../Drivers/STM32L4xx_HAL_Driver/Inc/stm32l4xx_hal_uart_ex.h \
+ ../Core/Inc/iwdg.h ../Core/Inc/stm32l476g_discovery_glass_lcd.h \
+ ../Core/Inc/stm32l476g_discovery.h
 ../Core/Inc/sai.h:
 ../Core/Inc/main.h:
 ../Drivers/STM32L4xx_HAL_Driver/Inc/stm32l4xx_hal.h:
@@ -73,3 +75,6 @@ Core/Src/sai.o: ../Core/Src/sai.c ../Core/Inc/sai.h ../Core/Inc/main.h \
 ../Drivers/STM32L4xx_HAL_Driver/Inc/stm32l4xx_hal_spi_ex.h:
 ../Drivers/STM32L4xx_HAL_Driver/Inc/stm32l4xx_hal_uart.h:
 ../Drivers/STM32L4xx_HAL_Driver/Inc/stm32l4xx_hal_uart_ex.h:
+../Core/Inc/iwdg.h:
+../Core/Inc/stm32l476g_discovery_glass_lcd.h:
+../Core/Inc/stm32l476g_discovery.h:
